Position vertices from tile-relative deltas, not absolute f32 ECEF
The vertex shaders computed absolute ECEF in f32 — a 0.5 m ulp at earth
radius — then subtracted the tile centre, snapping every vertex onto a
half-metre lattice that scalloped each straight edge. Tile uniforms now
carry the bounds relative to the centre and the centre's sin/cos (both
derived in double on the CPU), and the terrain and road shaders rebuild
each position as an ECEF delta from small, well-conditioned quantities
via exact trig identities, accurate to under a millimetre across a tile.
Bridge decks additionally draw through a small camera-facing depth margin
(the road paint's trick) so a deck coplanar with its engineered roadbed
shows its own smooth edge rather than a jagged depth-tie contour; tunnels
keep true depth and stay buried.

+++ client/src/tile/manager.c
@@ -880,15 +880,14 @@ static void draw_entry(arpt_renderer *r, const arpt_camera *cam,
                        const tile_entry *e) {
     arpt_mat4 model =
         arpt_camera_tile_model(cam, e->center_lon_rad, e->center_lat_rad, 0.0);
-    float bounds_rad[4] = {
-        (float)(e->bounds.west * M_PI / 180.0),
-        (float)(e->bounds.south * M_PI / 180.0),
-        (float)(e->bounds.east * M_PI / 180.0),
-        (float)(e->bounds.north * M_PI / 180.0),
+    double bounds_rad[4] = {
+        e->bounds.west * M_PI / 180.0,
+        e->bounds.south * M_PI / 180.0,
+        e->bounds.east * M_PI / 180.0,
+        e->bounds.north * M_PI / 180.0,
     };
     arpt_tile_gpu_set_uniforms((arpt_tile_gpu *)e->gpu, model, bounds_rad,
-                               (float)e->center_lon_rad,
-                               (float)e->center_lat_rad);
+                               e->center_lon_rad, e->center_lat_rad);
     arpt_renderer_draw_tile(r, (arpt_tile_gpu *)e->gpu);
 }
